building_footprint_api/requirements.txt
# Core web framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
starlette>=0.27.0
python-multipart>=0.0.6
email-validator>=2.0.0
fastapi-pagination>=0.12.4

# Database and caching
motor>=3.1.2
redis>=4.5.5
pymongo>=4.4.1

# Machine Learning
torch>=2.0.1
torchvision>=0.15.2
opencv-python>=4.7.0
scikit-learn>=1.2.2
scikit-image>=0.20.0
numpy>=1.24.3
scipy>=1.10.1
pillow>=9.5.0
matplotlib>=3.7.1
albumentations>=1.3.0

# Geospatial (managed by conda)
# gdal, rasterio, shapely, pyproj, geopandas installed via conda
geojson>=3.0.1

# Monitoring and logging
python-json-logger>=2.0.7
prometheus-client>=0.17.0
psutil>=5.9.5
sentry-sdk>=1.28.1

# WebSocket support
python-socketio>=5.8.0
websockets>=11.0.3

# API clients
httpx>=0.24.1
aiohttp>=3.8.5

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.2
rich>=13.4.2
pyyaml>=6.0
ujson>=5.8.0
typer>=0.9.0